requirements.txt
PyMySQL==1.1.0
PySimpleGUI==4.60.5
python-dotenv==1.0.0
